requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
pyyaml
streamlit
tensorflow